requirements.txt
python-telegram-bot==22.2
shazamio==0.8.1
faster-whisper==1.1.1
yt-dlp==2025.6.30
mysql-connector-python==8.2.0
pydub==0.25.1
python-dotenv==1.0.0
aiofiles==23.2.1
aiosqlite==0.19.0
python-multipart==0.0.6
ffmpeg-python==0.2.0
python-telegram-bot[ext]
